askrift-backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
sse-starlette==2.1.0
langchain==0.2.16
langchain-community==0.2.16
chromadb==0.5.5
sentence-transformers==3.0.1
groq==0.9.0
PyMuPDF==1.24.9
python-docx==1.1.2
supabase==2.7.2
bcrypt==4.2.0
pydantic-settings==2.4.0
python-dotenv==1.0.1
